==> 1_Geometry.ipynb <==
## Imports

import requests
import xyzservices.providers as xyz
from shapely import box

from utils import *

## US states & counties
Download the boundary shapefiles of the lower 48 US states and counties using the TIGER/Line API.

year = 2020

url = 'https://www2.census.gov/geo/tiger/TIGER{0}/STATE/tl_{0}_us_state.zip'.format(year)
states = (gpd.read_file(url).rename(columns=str.lower).to_crs(CRS_DEG)
          [['geoid', 'name', 'region', 'division', 'geometry']])
url = 'https://raw.githubusercontent.com/jasonong/List-of-US-States/refs/heads/master/states.csv'
state_abbr = pd.read_csv(url).rename(columns=D(State='name', Abbreviation='code'))
states = states.merge(state_abbr, on='name')
# rename US Census regions and divisions
states.region = states.region.astype(int).map({
    1: 'Northeast', 2: 'Midwest', 3: 'South', 4: 'West'})
states.division = states.division.astype(int).map({
    1: '1. New England', 2: '2. Middle Atlantic',
    3: '3. East North Central', 4: '4. West North Central',
    5: '5. South Atlantic', 6: '6. East South Central',
    7: '7. West South Central', 8: '8. Mountain', 9: '9. Pacific'
})
# remove Alaska and Hawaii
states = view(states[~states.name.isin(['Alaska', 'Hawaii'])]
              .sort_values('geoid').reset_index(drop=1)
              ['geoid code name region division geometry'.split()])
states.to_parquet('data/geometry/states.parquet') # 3s
# states = view(gpd.read_parquet('data/geometry/states.parquet'))

url = 'https://www2.census.gov/geo/tiger/TIGER{0}/COUNTY/tl_{0}_us_county.zip'.format(year)
counties = (gpd.read_file(url).rename(columns=str.lower)
            .rename(columns=D(name='county')).to_crs(CRS_DEG)
            ['geoid county aland awater geometry'.split()])
for x in ['aland', 'awater']: # convert area from sq.m. to sq.mi.
    counties[x] = (counties[x] * SQM2SQMI).astype(F32)
counties.to_parquet('data/geometry/counties.parquet') # 19s
# counties = view(gpd.read_parquet('data/geometry/counties.parquet'))

## Population
Download the population and labor force size of each census block group from the American Community Survey dataset and additionally aggregate to the tract and county levels.

popu_cols = D(B01001_001E='popu', B23025_002E='labor')
popu = []
for state in tqdm(states.geoid): # for each state
    url = (f'https://api.census.gov/data/{year}/acs/acs5?' +
           'get=' + ','.join(popu_cols) + '&for=block+group:*' +
           f'&in=state:{state}+county:*+tract:*')
    data = requests.get(url).json()
    df = Pdf(data[1:], columns=data[0])
    df.insert(0, 'geoid', df.pop('state') + df.pop('county') + 
              df.pop('tract') + df.pop('block group'))
    df = df.rename(columns=popu_cols).astype(D(popu=I32, labor=I32))
    popu.append(df.assign(level='BG'))
    for level, nchar in [('Tract', 11), ('County', 5)]:
        df2 = df.assign(geoid=df.geoid.str[:nchar])
        df2 = df2.groupby('geoid')[['popu', 'labor']].sum()
        popu.append(df2.reset_index().assign(level=level))
popu = view(pd.concat(popu).reset_index(drop=1).astype(D(level=CAT))
            [['geoid', 'level', 'popu', 'labor']])
popu.to_parquet('data/geometry/population.parquet') # 1m38s
# popu = view(pd.read_parquet('data/geometry/population.parquet'))

## Urban areas
Download the constituent census blocks of urban areas as defined by the Census Bureau. Then, identify the corresponding block groups and tracts which contain at least one block in each urban area.

url = f'https://www2.census.gov/geo/docs/reference/ua/{year}_UA_BLOCKS.txt'
cols = {'GEOID': 'geoid', f'{year}_UA_NAME': 'name', f'{year}_UACE': 'code'}
df = pd.read_csv(url, sep='|', encoding_errors='replace', usecols=list(cols))
df.rename(columns=cols, inplace=True)
df.code = df.code.astype(str).str.zfill(5)
df.geoid = df.geoid.astype(str).str.zfill(15)
urba = []
for level, nchar in [('County', 5), ('Tract', 11), ('BG', 12)]:
    df2 = df.assign(geoid=df.geoid.str[:nchar])
    df2 = df2.groupby(['code', 'name', 'geoid']).size()
    df2 = df2.rename('n').reset_index().sort_values('n', ascending=0)
    df2 = df2.drop_duplicates('geoid').sort_values('geoid')
    urba.append(df2.assign(level=level))
urba = pd.concat(urba).reset_index(drop=1)
urba = urba[['code', 'name', 'level', 'geoid']].astype(
    {x: CAT for x in ['level', 'name', 'code']})
urba.to_parquet('data/geometry/urban_zones.parquet') # 18s
# urba = view(pd.read_parquet('data/geometry/urban_zones.parquet'))

## 50 largest cities
Select the 50 most populous urban areas using the tract-level population.

def get_largest_cities(top, urba=urba, popu=popu):
    urba = urba.rename(columns=D(name='urba', code='ua_code'))
    df = urba.merge(popu[['geoid', 'popu']])
    df = df.groupby('urba', observed=False)['popu'].sum()
    df = df.sort_values(ascending=0).head(top).reset_index()
    df['city'] = [x.split(',')[0].split('-')[0].split('/')[0] for x in df.urba]
    print(Arr(df.city))
    df = urba[['geoid', 'level', 'ua_code', 'urba']].merge(df[['urba', 'city']])
    df = df.astype(D(level=str, urba=str, city=str))
    return df

top50 = view(get_largest_cities(50))

## Zones
Download the shapefile of the block groups and tracts which lie in the largest urban areas. Remove the zones with no land area for better plotting.

uri = f'https://www2.census.gov/geo/tiger/TIGER{year}'
cols = 'geoid level county ua_code city aland awater popu labor geometry'.split()
zones = (counties.merge(top50, on='geoid')
         .merge(popu.drop(columns='level'), on='geoid')[cols])
# for each state which includes at least one major urban area
for state in tqdm(set(top50.geoid.str[:2])):
    for level in ['bg', 'tract']:
        df = gpd.read_file('{}/{}/tl_{}_{}_{}.zip'.format(
            uri, level.upper(), year, state, level))
        df = df[df.ALAND > 0] # remove water-only zones
        df = df.to_crs(CRS_DEG).rename(columns=str.lower)
        for x in ['aland', 'awater']: # convert area from sq.m. to sq.mi.
            df[x] = (df[x] * SQM2SQMI).astype(F32)
        df = df.merge(top50, on='geoid')
        df = df.merge(popu.drop(columns='level'), on='geoid')
        df['county_id'] = df.geoid.str[:5]
        df = df.merge(counties.rename(columns=D(geoid='county_id'))
                      [['county_id', 'county']], on='county_id')
        zones = pd.concat([zones, df[cols]], ignore_index=True)
zones = view(zones.reset_index(drop=1)
             .astype(D(level=CAT, county=CAT, ua_code=CAT, city=CAT)))
zones.to_parquet('data/geometry/zones.parquet') # 5m37s
# zones = view(gpd.read_parquet('data/geometry/zones.parquet'))

## Zone centroids
Coordinates of zone centroids to be used by the OSRM routing engine in `get_ttm.py` (via `osrm.sh`).

pts = (zones.set_index(['geoid', 'level', 'city']).to_crs(CRS_M)
       .centroid.to_crs(CRS_DEG).get_coordinates().reset_index()
       .astype(D(level=str, city=str, x=F32, y=F32)))
pts.city = [x.lower().replace(' ', '-').replace('.', '') for x in pts.city]
pts.level = pts.level.str.lower()
view(pts).to_parquet('data/geometry/centroids.parquet') # 6s
# pts = view(pd.read_parquet('data/geometry/centroids.parquet'))

## City to region
Associate each city with a US Census region and find its bounding box using its block group (BG) geometry. This is to be used by `extract_osm.sh` to prepare the OSM extracts of the urban areas using the extracts of the US Census regions. For cities spanning multiple regions, select the region with the most BGs.

bounds = (zones.set_index('city').bounds
          .groupby('city', observed=1).agg(D(
              minx='min', miny='min', maxx='max', maxy='max'))
          .apply(lambda x: np.round(x, 6)))
city2rgn = (filt(zones, level='BG')
            .assign(geoid=lambda df: df.geoid.str[:2])
            .groupby(['geoid', 'city'], observed=1)
            .size().rename('nBG').sort_values(ascending=0)
            .reset_index().drop_duplicates(subset='city')
            .merge(states, on='geoid')[['city', 'region']]
            .merge(bounds, on='city'))
for col in ['city', 'region']:
    city2rgn[col] = [x.lower().replace(' ', '-').replace('.', '')
                     for x in city2rgn[col]]
view(city2rgn, None).to_csv('data/geometry/city_region.csv', index=False)

## Fig: City map

top_cities = view(filt(zones, level='Tract')[['city', 'popu', 'geometry']]
                  .dissolve('city', aggfunc='sum').to_crs(CRS_M)
                  .assign(rank_=lambda df: df.popu.rank(ascending=0))
                  .astype(D(rank_=I16)).reset_index()) # 25s

xyz.MapBox['accessToken'] = 'pk.eyJ1IjoiZW1waGFzZW50IiwiYSI6ImNtMGgyN3dqcTA2eHYyanBzeWUxZWc4YnYifQ.f6gyGE1jH2vijK9tHPVQMQ'

# ax = plot(size=(16, 8), dpi=250, axoff=1)
# df = filt(zones, level='County').to_crs(CRS_M)
# df.plot(ax=ax, fc='yellow', ec='k', lw=0.2)
# df = filt(zones, level='Tract').to_crs(CRS_M)
# df.plot(ax=ax, color='tomato', alpha=1)
# states.to_crs(CRS_M).plot(ax=ax, fc='none', ec='k', lw=0.25)
# ctx.add_basemap(ax=ax, crs=CRS_M, source=ctx.providers.MapBox, alpha=0.6)
# for _, r in top_cities.iterrows():
#     ax.text(*r.geometry.centroid.coords[0], str(r.rank_),
#             ha='center', va='center', fontsize=9, color='w',
#             bbox=D(boxstyle='circle', pad=0.1, fc='k', alpha=0.6))
# plt.savefig('fig/study_cities_map.png', dpi=300) # 23s

## Map of each city

city_limits = view(filt(zones, level='BG').dissolve('city')
                   .reset_index().to_crs(CRS_M), 0) # 47s

counties = view(filt(zones, level='County').to_crs(CRS_M), 0)

# total 4m39s
city_ranks = (filt(zones, level='BG').groupby('city', observed=1)
              .popu.sum().sort_values(ascending=0).reset_index()
              .assign(rank=lambda df: df.index + 1)
              .set_index('city')['rank'])
for batch, cities in enumerate(np.split(city_ranks.index, 2)):
    fig, axes = plt.subplots(5, 5, figsize=(30, 30), dpi=100)
    for rank, city in enumerate(pbar := tqdm(cities)):
        pbar.set_description(city)
        cnty = filt(zones, city=city, level='County').to_crs(CRS_M)
        boundary = filt(city_limits, city=city)
        minx, miny, maxx, maxy = boundary.bounds.iloc[0]
        cx, cy = (maxx + minx) / 2, (maxy + miny) / 2
        width, height = maxx - minx, maxy - miny
        d = max(width, height) / 2 # half-width of the desired square
        bbox = x0, y0, x1, y1 = cx - d, cy - d, cx + d, cy + d
        bbox = Gdf(geometry=[box(*bbox)], crs=CRS_M)
        ax = axes.flatten()[rank]
        ax.set_title(f'{city_ranks.loc[city]}. {city}',
                     fontsize=20, fontweight='bold')
        boundary.plot(ax=ax, color='tomato', ec='none', alpha=0.4)
        cnty.plot(ax=ax, fc='none', ec='k')
        for _, r in cnty.iterrows():
            x, y = r.geometry.centroid.coords[0]
            if x < x0 or x > x1 or y < y0 or y > y1:
                continue
            ax.text(x, y, r.county, fontsize=10, ha='center', va='center')
        bbox.plot(ax=ax, alpha=0)
        basemap(ax, crs=CRS_M)
        ax.set_xlim(x0, x1)
        ax.set_ylim(y0, y1)
        # break
    plt.subplots_adjust(wspace=0.05, hspace=0.1)
    imsave(f'City boundaries {batch + 1}', fig=fig)
    plt.close()
