==> 2. Opportunities.ipynb <==
## Setup

from utils import *

## Zones

bgs = gpd.read_parquet('data/zones.parquet', filters=[
    ('level', '==', 'BG')], columns=['geoid', 'geometry']).disp()

## Job counts

Download LODES data (https://lehd.ces.census.gov/data) for states which contain at least one CBG of the 50 largest cities.

imp_states = (pd.read_parquet('data/states.parquet', columns=['geoid', 'code'])
              .merge(bgs.geoid.str[:2], on='geoid').drop_duplicates().code)
Arr(imp_states)

Download the job count data for 2021 using the Workplace Area Characteristics (WAC) table of the LODES v8 dataset. For Arizona (AR) and Mississippi (MS), the latest data is available for 2018 (for which the dataset version is v7).

# jobs = []
# for abbr in tqdm(imp_states):
#     # download data for 2021 except for AL & MS (outdated data)
#     year = 2018 if abbr in ['AR', 'MS'] else 2021
#     version = 7 if year <= 2020 else 8
#     cols = D(w_geocode='geoid', C000='All', CE01='Low wage', CR01='White',
#          CD01='< high school', CD02='high school')
#     url = (f'https://lehd.ces.census.gov/data/lodes/LODES{version}/' + 
#            '{0}/wac/{0}_wac_S000_JT00_{1}.csv.gz'.format(abbr.lower(), year))
#     df = pd.read_csv(url, usecols=list(cols)).rename(columns=cols)
#     df['Low edu'] = df.pop('< high school') + df.pop('high school')
#     df['POC'] = df['All'] - df.pop('White')
#     df.geoid = df.geoid.astype(str).str.zfill(15)
#     for level, nchar in [('BG', 12), ('Tract', 11), ('County', 5)]:
#         df2 = df.assign(geoid=df.geoid.str[:nchar])
#         df2 = df2.groupby('geoid').sum().reset_index()
#         df2 = df2.melt('geoid', var_name='kind', value_name='njobs')
#         jobs.append(df2.query('njobs > 0').assign(level=level))
# cols = D(geoid=CAT, level=CAT, kind=CAT, njobs=I32)
# jobs = pd.concat(jobs).reset_index(drop=1).astype(cols)[list(cols)].disp()
# jobs.to_parquet('data/jobs.parquet') # 50s
jobs = pd.read_parquet('data/jobs.parquet').disp()

## POIs
Process the Points of Interest (POI) data from the SafeGraph dataset of Nov 2020 obtained previously as part of the open research data sharing contract during Covid.

poi_cats = Pdf(sum([[(k, x) for x in v] for k, v in {
    'Education': [
        'Colleges, Universities, and Professional Schools',
        'Elementary and Secondary Schools',
        'Other Schools and Instruction',
    ],
    'Groceries': [
        'Department Stores',
        'General Merchandise Stores, Including Warehouse Clubs and Supercenters',
        'Grocery Stores',
    ],
    'Medical': [
        'General Medical and Surgical Hospitals',
        'Medical and Diagnostic Laboratories',
        'Nursing Care Facilities (Skilled Nursing Facilities)',
        'Offices of Dentists',
        'Offices of Other Health Practitioners',
        'Offices of Physicians',
        'Other Ambulatory Health Care Services',
        'Outpatient Care Centers',
        'Specialty (except Psychiatric and Substance Abuse) Hospitals',
    ],
    'Social Support': [
        'Child Day Care Services',
        'Community Food and Housing, and Emergency and Other Relief Services',
        'Continuing Care Retirement Communities and Assisted Living Facilities for the Elderly',
        'Home Health Care Services',
    ]
}.items()], []), columns=['kind', 'top_category']).disp()

# pois = pd.read_parquet(
#     '/home/umni2/a/umnilab/data/safegraph/pois/us/2020-11-06.parquet',
#     columns=['top_category', 'lon', 'lat'])
# pois = (pdf2gdf(pois, crs=CRS_DEG)
#         .sjoin(bgs, predicate='within')
#         .groupby(['geoid', 'top_category']).size()
#         .rename('npois').reset_index()
#         .merge(poi_cats, 'left', on='top_category'))
# bg_pois = pd.concat([
#     pois.groupby('geoid').npois.sum().reset_index().assign(kind='All'),
#     pois.groupby(['kind', 'geoid']).npois.sum().reset_index()
# ]).reset_index(drop=1).assign(level='BG')
# pois = [bg_pois]
# for level, nchar in [('Tract', 11), ('County', 5)]:
#     df = bg_pois.assign(geoid=bg_pois.geoid.str[:nchar])
#     df = df.groupby(['geoid', 'kind']).npois.sum().reset_index()
#     pois.append(df.assign(level=level))
# pois = pd.concat(pois).reset_index(drop=1)
# cols = D(geoid=CAT, level=CAT, kind=CAT, npois=I32)
# pois = pois.astype(cols)[list(cols)].disp()
# pois.to_parquet('data/pois.parquet') # 21s
pois = pd.read_parquet('data/pois.parquet').disp()

filt(pois, level='BG').groupby('kind', observed=0).npois.sum().to_frame().T

## Combine job and POI counts

# opport = pd.concat([
#     jobs.assign(kind='Jobs: ' + jobs.kind.astype(str)).rename(columns=D(njobs='opport')),
#     pois.assign(kind='POIs: ' + pois.kind.astype(str)).rename(columns=D(npois='opport'))
# ]).reset_index(drop=1).astype(D(geoid=CAT, kind=CAT)).disp()
# opport.to_parquet('data/opportunities.parquet')
opport = pd.read_parquet('data/opportunities.parquet').disp()

## City summary statistics

bgs = (
    pd.read_parquet('data/zones.parquet', filters=[('level', '==', 'BG')])
    .merge(filt(jobs, kind='All', level='BG'), 'left')
    .merge(filt(pois, kind='All', level='BG'), 'left')
    .astype(D(county=str, ua_code=str, city=str))
    .drop(columns='geometry')
).disp()

df = bgs.groupby(['ua_code', 'city']).agg(D(
    geoid='count', county=set, aland='sum', awater='sum',
    popu='sum', njobs='sum', npois='sum'
)).reset_index().rename(columns=D(geoid='nBG'))
ntract = (bgs.assign(geoid=bgs.geoid.str[:11])
          .groupby(['city', 'geoid']).size().reset_index()
          .groupby('city').geoid.count().rename('nTract').reset_index())
df = df.merge(ntract, on='city')
df['popdens'] = df.popu / df.aland
df['nCounty'] = df.county.apply(len)
df.index = df.popu.rank(ascending=0).astype(I16).rename('rank')
for col in ['popu', 'njobs', 'npois']:
    df[col] /= 1000
df = df['city ua_code nBG nTract nCounty popu aland popdens njobs npois'.split()]
df = df.sort_index().disp(None)
df.to_csv('data/city_summary_stats.csv')

df = pd.read_csv('data/export/las-vegas/las-vegas_bg.csv').disp()
